# inventarioBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Inventario API",
        "description": "Sistema para gestionar inventario para empresas en crecimiento",
        "contact": {
            "name": "API Support",
            "email": "support@inventario.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/facturas": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Listar facturas con detalles",
                "description": "Listar todas las facturas con sus detalles completos",
                "operationId": "93b68ab45f377135f06da3bd8429498f",
                "parameters": [
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "cliente",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "fecha"
                        }
                    },
                    {
                        "name": "order_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de facturas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Crear factura con detalles",
                "description": "Crear una nueva factura con sus detalles",
                "operationId": "9865ea9ed3a6ae3a571ef92e00823c32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "detalles"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_facturas": {
                                        "type": "string",
                                        "example": "F-000123",
                                        "nullable": true
                                    },
                                    "detalles": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "producto_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cantidad": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Factura creada"
                    }
                }
            }
        },
        "/facturas/{id}": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Mostrar factura con detalles",
                "description": "Mostrar una factura específica con todos sus detalles",
                "operationId": "8ee04f9a6e8876fd71ce7c32085d2763",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Factura encontrada"
                    },
                    "404": {
                        "description": "Factura no encontrada"
                    }
                }
            }
        },
        "/facturas/{id}/detalles": {
            "post": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Agregar detalle a factura",
                "description": "Agregar un detalle a una factura existente\nPayload: factura_id, producto_id, cantidad",
                "operationId": "ace3081a23f2398a066842af1846085b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "factura_id",
                                    "producto_id",
                                    "cantidad"
                                ],
                                "properties": {
                                    "factura_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "producto_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detalle agregado"
                    }
                }
            }
        },
        "/facturas/detalles/{id}": {
            "put": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Actualizar detalle de factura",
                "description": "Actualizar un detalle (p. ej. cambiar cantidad)",
                "operationId": "20ac4c53ab2066b271b56f68ea2ddcf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detalle actualizado"
                    },
                    "404": {
                        "description": "Detalle no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Eliminar detalle de factura",
                "description": "Eliminar un detalle de factura y restaurar stock",
                "operationId": "b9464ee6d8bf6be68af22fddf2d50353",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle eliminado"
                    },
                    "404": {
                        "description": "Detalle no encontrado"
                    }
                }
            }
        },
        "/facturas/productos/listar": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Listar productos para facturación",
                "description": "Listar productos disponibles (paginado, con búsqueda)",
                "operationId": "883d405edb0e29c6dd901b145870817e",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos"
                    }
                }
            }
        },
        "/facturas/productos/{id}": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Mostrar producto",
                "description": "Mostrar un producto específico",
                "operationId": "06da958e6993e8c8791e20ed1effb592",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
        "/facturas/clientes/listar": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Listar clientes para facturación",
                "description": "Listar clientes (útil para seleccionar cliente desde Angular)",
                "operationId": "7d1778db99c30dbc02ca031b64d724c6",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "name": "bearerAuth",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}